requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
openai==1.99.5
pydantic==2.11.7
pydantic-core==2.33.2
sniffio==1.3.1
tqdm==4.67.1
typing-extensions==4.14.1
typing-inspection==0.4.1
# Japanese Grammar Correction project dependencies
mlx>=0.20.0
mlx-lm>=0.20.0
numpy>=1.24.0
pyyaml>=6.0
transformers>=4.30.0
tokenizers>=0.13.0
datasets>=2.10.0